diff cleanup
